LayoutTests/http/tests/xmlhttprequest/access-control-preflight-credential-async-expected.txt

ALERT: http://localhost:8000/cookies/resources/cookie-utility.php?queryfunction=setFooCookie
ALERT: XHR response - Set the foo cookie
Test case for bug 37781: [XHR] Cross-Origin asynchronous request with credential raises NETWORK_ERR

PASSED
